+++ components/rlc_common/include/rlc_version.h
@@ -8,6 +8,6 @@
 #pragma once
 
 #define RLC_VERSION_MAJOR  1
-#define RLC_VERSION_MINOR  0
+#define RLC_VERSION_MINOR  1
 #define RLC_VERSION_PATCH  0
-#define RLC_VERSION_STRING "1.0.0"
+#define RLC_VERSION_STRING "1.1.0"